added alert to delete
in addition to that single thing i learned a lot, if you read this I
think I need help mainly writing to the db and local, logic when we
delete locally if it changes the bd.

Let me struggle through the rest of the UI stuff, in layout. you can
start adding images to icons once i implement the menu tomorrow
hopefully, i learned how to tonight.