kept going on the tutorial
I didnt start at the beginning of the tutorial and just added to ours, o
i missed something, will have to go back to see.

added second view details of the todo activity fragment and activity___
it needs to ling layout to fragment through activity

also finished custom array adapters for the list of todo
formatted<string arrays>

made the view layout in the new fragment, we can at least salvage the
fragment and custom class files if we have to start anew on the main
activity so they feed layouts correctly.